Add a specific page to the feed, by hand
I don't know how to do it in pure Jekyll, so i've cheated.

To generate the "feed" version for this post, I have used
`site.pages limit:999` for the loop, and
`post.content | markdownify` to convert the content to HTML.

The I edited the results, isolating the desired page and converting all
relative links and image paths to full URLs.