New _scripts/generate-index to create missing index.html (#19)
Apache used to list folder contents automatically.
In Netlify this does not happen, so I have to create my own
index.html files via a script.

I could have done it in Python, but the shell version got simple
enough.

I could have done the HTML entirely in the script. But I preferred
to only set Front Matter properties and let Jekyll and Liquid do
the job of interating and filling the template.

The downside is that I'll have to port or rewrite the solution if
I move to another static site generator.