regex/guia: Remove layout HTML, leave content only (#66)
This is the first step of moving all this content to a Jekyll
template instead of hardcoded layout HTML in all files.

In lots of sucessive replaces, I've removed headers, navigation, ads,
etc until all that has left is the page contents.

In this step I'm also setting properly in the Front Matter the
page title and prev/next book navigation.

For now I'm only dealing with the HTML pages themselves. The config
and template files will come in a later commit. It's too much work
to do in a single step.