regex/guia: Remove silly empty topics (#66)
My idea in the past was to call attention to the new topics
available in the new book. But that's silly. They are different
books, no need to cross contents like that.

This is the online version for the 2001 book. That's it.

I've removed all the extra files and rewired the prev/next chain.
I've set up Gone redirects for all the old URLs.