Pin redcarpet version to 3.3.4
This was the last version they are correctly using &amp; to escape &
inside `<a href>` URLs.

    -<a href="https://example.com/?foo=1&amp;bar=2">Foo</a>
    +<a href="https://example.com/?foo=1&bar=2">Foo</a>

BTW, redcarpet is not supported anymore in Jekyll 4.x. For the future,
I'll have to decide what to do. See issue #71.

Since I've updated some gems versions, the generated HTML code changed
a little on the syntax highlight for code blocks (due to new rouge
version).

As far as I saw, it did not brake anything noticeable.