_scripts/check-site: New check_broken_html() (#77)
The idea here is having a generic checker for common (critical)
mistakes when editing HTML files by hand.

For now, it only checks for broken link tags (`<a>`). It's sad to have
a broken link. But having a broken link tag is a total failure.

All the offending HTML files were also fixed in this commit, by using:

    replace -i -r -f '(href="[^"]+?)>' -t '\1">' **.html