Bump versions for Ruby and Bundler
This is the result of running `bundle install` after I did the
following updates:

- Ruby version: 2.4.3 -> 3.0.2
- Bundler version: 2.0.1 -> 2.3.5

Those versions now match what's inside the devcontainer (see previous
commit), so from now on we can have everything pinned and stable.

Jekyll and redcarpet versions were not bumped.

The only noticeable difference in the generated HTML was the use of
`&apos;` instead of a literal ', when appearing in an escaped text
(see `feed.xml` file).

    '\w+' -> &apos;\w+&apos;